pyproject.toml
[project]
name = "spacexdata"
version = "0.1.0"
description = "GraphQL wrapper for SpaceX API"
requires-python = ">=3.9"
dependencies = [
    "strawberry-graphql>=0.219.2",
    "httpx>=0.26.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build" 